Register AssetInfo to IObject allowing you to get the object which is related to the asset on disk.

+++ Engine/Editor/src/Editor/EditorWindows/AssetAuditWindow.cpp
@@ -77,7 +77,7 @@ namespace Insight
                 }
 
                 ImGuiListClipper clipper;
-                clipper.Begin(assetInfos.size());
+                clipper.Begin(static_cast<int>(assetInfos.size()));
                 
                 while (clipper.Step())
                 {
